feat(toolbox): separated web from native HangupButton
