+++ react/features/toolbox/components/index.js
@@ -2,5 +2,5 @@
 
 export * from './_';
 export { default as AudioMuteButton } from './AudioMuteButton';
-export { default as HangupButton } from './HangupButton';
+export { default as HangupButton } from './web/HangupButton';
 export { default as VideoMuteButton } from './VideoMuteButton';
