+++ react/features/toolbox/components/native/Toolbox.js
@@ -14,7 +14,7 @@ import { isReactionsEnabled } from '../../../reactions/functions.any';
 import { TileViewButton } from '../../../video-layout';
 import { getMovableButtons, isToolboxVisible } from '../../functions.native';
 import AudioMuteButton from '../AudioMuteButton';
-import HangupButton from '../HangupButton';
+import HangupButton from '../native/HangupButton';
 import VideoMuteButton from '../VideoMuteButton';
 
 import HangupMenuButton from './HangupMenuButton';
