+++ react/features/toolbox/components/web/HangupButton.js
@@ -1,0 +1,63 @@
+// @flow
+
+import _ from 'lodash';
+
+import { createToolbarEvent, sendAnalytics } from '../../../analytics';
+import { disconnect } from '../../../base/connection';
+import { translate } from '../../../base/i18n';
+import { connect } from '../../../base/redux';
+import { AbstractHangupButton } from '../../../base/toolbox/components';
+import type { AbstractButtonProps } from '../../../base/toolbox/components';
+
+
+/**
+ * The type of the React {@code Component} props of {@link HangupButton}.
+ */
+type Props = AbstractButtonProps & {
+
+    /**
+     * The redux {@code dispatch} function.
+     */
+    dispatch: Function
+};
+
+/**
+ * Component that renders a toolbar button for leaving the current conference.
+ *
+ * @augments AbstractHangupButton
+ */
+class HangupButton extends AbstractHangupButton<Props, *> {
+    _hangup: Function;
+
+    accessibilityLabel = 'toolbar.accessibilityLabel.hangup';
+    label = 'toolbar.hangup';
+    tooltip = 'toolbar.hangup';
+
+    /**
+     * Initializes a new HangupButton instance.
+     *
+     * @param {Props} props - The read-only properties with which the new
+     * instance is to be initialized.
+     */
+    constructor(props: Props) {
+        super(props);
+
+        this._hangup = _.once(() => {
+            sendAnalytics(createToolbarEvent('hangup'));
+            this.props.dispatch(disconnect());
+        });
+    }
+
+    /**
+     * Helper function to perform the actual hangup action.
+     *
+     * @override
+     * @protected
+     * @returns {void}
+     */
+    _doHangup() {
+        this._hangup();
+    }
+}
+
+export default translate(connect()(HangupButton));
